% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.compareToHealthy.R
\name{compareToHealthy}
\alias{compareToHealthy}
\alias{compareToHealthy,fragSet-method}
\title{Compare sample fragmentation profile to healthy median}
\usage{
compareToHealthy(object, healthy)

\S4method{compareToHealthy}{fragSet}(object, healthy)
}
\arguments{
\item{object}{fragSet object}

\item{healthy}{Character of path to healthy median rda file}
}
\value{
Returns a fragSet object with mb5dist and md5summary dataframes
stored in results slot, access with get5MbDist() and get5MbSummary()
respectively
}
\description{
Calculate correlation and distance of sample fragmentation profile to
healthy median (provided with package or generated externally).
}
