% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.fragSetClass_createFragSet.R
\name{createFragSet}
\alias{createFragSet}
\title{Initiate fragSet object}
\usage{
createFragSet(
  sampleID,
  filePath,
  fileType,
  refGenome = "BSgenome.Hsapiens.UCSC.hg38",
  chr.select = paste0("chr", 1:22),
  width_min = 90,
  width_max = 220,
  mapQthreshold = 30,
  keepDuplicates = FALSE,
  keepSecondaryAlignments = FALSE
)
}
\arguments{
\item{sampleID}{Character of sample ID}

\item{filePath}{Character of full path to input file}

\item{fileType}{Character of either "bam" or "bedpe"}

\item{refGenome}{Character of reference genome used (default is BSgenome.Hsapiens.UCSC.hg38)}

\item{chr.select}{Vector of chromosome names (default is paste0("chr",1:22))}

\item{width_min}{Numeric of minimum fragment length (default is 90)}

\item{width_max}{Numeric of maximum fragment length (default is 220)}

\item{mapQthreshold}{Numeric of mapping quality filter greater or equal to value (default is 30)}

\item{keepDuplicates}{Boolean of whether to keep duplicates reads (default is FALSE - remove duplicate reads with FLAG 0x400 set)}

\item{keepSecondaryAlignments}{Boolean of whether to keep secondary reads (default is FALSE - remove secondary reads with FLAG 0x100 set)}
}
\value{
Returns a fragSet object
}
\description{
createFragSet() will read in BAM or BEDPE input file for a sample,
filter reads based on SAM FLAG, fragment width, and mapQ threshold. Then
stores all fragments in a GRanges object with GC content information from
the reference genome specified. This frags GRanges object is stored in the
"frags" slot. Fragments mapping to mitochondrial genome is stored in a
separate GRanges object and stored in the "mito" slot.
}
