% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.fragSetClass_createFragSet.R
\docType{class}
\name{fragSet-class}
\alias{fragSet-class}
\title{The fragSet Class}
\description{
The fragSet object is a storage container used to store
all intermediate and final results of fragmentomics analysis
}
\section{Slots}{

\describe{
\item{\code{sampleID}}{Character of sample ID}

\item{\code{refGenome}}{Character of reference genome used (default is BSgenome.Hsapiens.UCSC.hg38)}

\item{\code{inputFileType}}{Character of input file type ("bam" or "bedpe")}

\item{\code{frags}}{List to store non-mitochondrial fragments GRanges object}

\item{\code{galp_len}}{Numeric of the length of genomic alignment pair vector (i.e. total number of fragments sequenced)}

\item{\code{mito}}{List to store mitochondrial fragments GRanges object, and mitochondrial fragment report}

\item{\code{bins}}{List to store 100kb bins and 5Mb bins results}

\item{\code{results}}{List to store 5Mb distance to healthy median, 5Mb correlation to healthy median, fragmentomics profile plots}

\item{\code{QC}}{List to store high coverage regions, coverage statistics and GC metrics}
}}

