% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.gen5MbBins.R
\name{gen5MbBins}
\alias{gen5MbBins}
\alias{gen5MbBins,fragSet-method}
\title{Generate 5Mb bin fragment coverage dataframe from 100kb bin dataframe}
\usage{
gen5MbBins(object)

\S4method{gen5MbBins}{fragSet}(object)
}
\arguments{
\item{object}{fragSet object}
}
\value{
Returns a fragSet object with 5Mb bin fragment coverage dataframe stored
in "bins" slot, access with get5MbBins() function
}
\description{
The 5Mb bin fragment coverage dataframe including the following columns:
"sample_id", "seqnames", "arm", "start", "end", "gc", "frag_gc",
"short", "long", "nfrags", "ratio", "short_corrected", "long_corrected",
"nfrags_corrected", "ratio_corrected", "ratio_centered", "coverage",
"coverage_corrected", "coverage_centered", "combined", "combined_centered",
"short_var", "long_var", "nfrags_var", "mode_size", "mean_size",
"median_size", "q25_size", "q75_size", "binsize", "bin", "combine"
}
