% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.genMitoReport.R
\name{genMitoReport}
\alias{genMitoReport}
\alias{genMitoReport,fragSet-method}
\title{Generate mitochondrial fragment report}
\usage{
genMitoReport(object)

\S4method{genMitoReport}{fragSet}(object)
}
\arguments{
\item{object}{fragSet object}
}
\value{
Returns a fragSet object with mitochondrial fragment report stored in "mito" slot, access with getMitoReport() function
}
\description{
Mitochondrial fragment report includes: mt_short (number of mito short reads),
mt_long (number of mito long reads), mt_nFrag (total number of fragments mapped to mito),
mt_ratio (ratio of number of short/long mito reads), mt_coverage (sum(mt_width)/16569),
mt_fraction (number of mitochondrial fragments / total number of fragments),
mt_median/mean/mode/lower_quartile/upper_quartile
(median/mean/mode/lower and upper quartile of mitochondrial fragment distribution),
and sampleID
}
