% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.writeOutputAndPlots.R
\name{outputAllResults}
\alias{outputAllResults}
\alias{outputAllResults,fragSet-method}
\title{Write out dataframe results as text files}
\usage{
outputAllResults(object, outdir)

\S4method{outputAllResults}{fragSet}(object, outdir)
}
\arguments{
\item{object}{fragSet object}

\item{outdir}{Character of full path to output directory}
}
\value{
None
}
\description{
Write out mitochondrial report, high coverage regions, coverage statistics,
100kb bin, 5Mb bin, 5Mb distance to healthy median, 5Mb summary files.
}
