% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.writeOutputAndPlots.R
\name{outputFragPlots}
\alias{outputFragPlots}
\alias{outputFragPlots,fragSet-method}
\title{Write out plots related to fragmentation profile}
\usage{
outputFragPlots(object, healthy, outdir)

\S4method{outputFragPlots}{fragSet}(object, healthy, outdir)
}
\arguments{
\item{object}{fragSet object}

\item{healthy}{Character of path to healthy median rda file}

\item{outdir}{Character of full path to output directory}
}
\value{
None
}
\description{
Write out plots including ...fragment.pdf (short/long fragment ratio in 5Mb bins),
...coverage.pdf (short fragment coverage in 5Mb bins),
...coverage.pdf (ratio*coverage in 5Mb bins).
}
