% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag.writeOutputAndPlots.R
\name{outputGCcorrectPlots}
\alias{outputGCcorrectPlots}
\alias{outputGCcorrectPlots,fragSet-method}
\title{Write out plots related to GC bias correction}
\usage{
outputGCcorrectPlots(object, outdir)

\S4method{outputGCcorrectPlots}{fragSet}(object, outdir)
}
\arguments{
\item{object}{fragSet object}

\item{outdir}{Character of full path to output directory}
}
\value{
None
}
\description{
Write out plots before and after GC bias correction
}
